+++ CHANGELOG.md
@@ -6,6 +6,7 @@ FEATURES:
 * add `wallix-bastion_externalauth_saml` resource
 * add `wallix-bastion_authdomain_ldap` resource
 * add `wallix-bastion_authdomain_mapping` resource
+* add `wallix-bastion_authdomain_ad` resource
 
 ENHANCEMENTS: